requirements.txt
mcp>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
httpx>=0.27.0
openai>=1.3.1
python-dotenv>=1.0.0
gradio>=4.0.0
